# swagger.yaml
openapi: 3.0.0
info:
  title: Swagger API project
  description: learning swagger.
  version: 0.1.9
  contact:
    email: abinjola.binjola@gmail.com
    url: "https://ankitbinjola.com"

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for local server
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http
          

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /lco:
    get:
      tags:
        - String
      summary: returns a greeting msg from lco
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
                example: "mystring"
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /lcoobject:
    get:
      tags:
        - Object
      summary: returns an Object containing values
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /courses:
    get:
      tags:
        - Array
      summary: returns an array of Objects containing values
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /courses/{CourseId}:
    get:
      tags:
        - String
      summary: returns an  Objects based on Id params
      parameters:
        - name: CourseId
          in: path
          description: Course Id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error         


  /addcourse:
    post:
      tags:
        - String
      summary: returns boolean value after adding course
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true 
        description: A json object 
        content: 
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: boolean
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error         

  /coursequery:
    get:
      tags:
        - String
      summary: returns querrt params
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [delhi, london, kotdwar]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [mobile, web]       
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error         


  /uploadfile:
    post:
      tags:
        - String
      summary: returns uploading course image
      parameters:
        - in: header
          name: auth
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error